+++ Cargo.toml
@@ -24,6 +24,9 @@ wasm = [
   "console_error_panic_hook", "chrono/wasmbind"
 ]
 
+[build-dependencies]
+protobuf-codegen = "3.4.0"
+protoc-bin-vendored = "3.0.0"
 
 [dependencies]
 test-case = "3.2.1"
@@ -34,6 +37,7 @@ serde_json = { version = "1.0.108", features = ["raw_value"] }
 chrono = { version = "0.4.31", features = ["serde"] }
 duration-str = "0.7.0"
 num = "0.4.1"
+protobuf = { version = "3.4.0" }
 
 # Dependencies for python bindings
 pyo3 = { version = "0.20.0", optional = true, features = ["extension-module", "chrono"] }
