+++ src/tests/mod.rs
@@ -1,0 +1,3 @@
+mod general_tests;
+mod proto_tests;
+mod type_prop_tests;
